checkpoint: created Publisher core; DependencyProvider inherits from it
