+++ nameko/amqp/__init__.py
@@ -9,6 +9,10 @@ from kombu import Connection
 from kombu.pools import connections, producers
 from kombu.transport.pyamqp import Transport
 
+# backwards compat imports
+from .publish import get_connection, get_producer, UndeliverableMessage
+
+
 BAD_CREDENTIALS = (
     'Error connecting to broker, probably caused by invalid credentials'
 )
@@ -50,27 +54,3 @@ def verify_amqp_uri(amqp_uri):
     transport = TestTransport(connection.transport.client)
     with transport.establish_connection():
         pass
-
-
-class UndeliverableMessage(Exception):
-    """ Raised when publisher confirms are enabled and a message could not
-    be routed or persisted """
-    pass
-
-
-@contextmanager
-def get_connection(amqp_uri):
-    conn = Connection(amqp_uri)
-    with connections[conn].acquire(block=True) as connection:
-        yield connection
-
-
-@contextmanager
-def get_producer(amqp_uri, confirms=True):
-    transport_options = {
-        'confirm_publish': confirms
-    }
-    conn = Connection(amqp_uri, transport_options=transport_options)
-
-    with producers[conn].acquire(block=True) as producer:
-        yield producer
